Youtube (music) api at yt_api.js
